JS15/TP_JS15/requirements.txt
flask
gunicorn
torch
torchvision
Pillow
numpy
scikit-learn
scikit-image
opencv-python-headless